requirements.txt
# Core ML and Deep Learning
tensorflow>=2.12.0
tensorflow-addons>=0.20.0
keras>=2.12.0
scikit-learn>=1.2.0
numpy>=1.21.0
pandas>=1.5.0

# Image Processing
opencv-python>=4.7.0
Pillow>=9.0.0
albumentations>=1.3.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Data Management
kaggle>=1.5.0
tqdm>=4.64.0
h5py>=3.7.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
pathlib>=1.0.0

# Metrics and Evaluation
scikit-metrics>=0.1.0
tensorboard>=2.12.0

# Model Deployment
flask>=2.2.0
gunicorn>=20.1.0

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=4.0.0
jupyter>=1.0.0
notebook>=6.4.0

# Optional GPU support
# tensorflow-gpu>=2.12.0  # Uncomment if using GPU